requirements.txt
fastapi[all]
python-dotenv
pg8000
pytest